Project state at end of HACS408T Pentesting class
This is the state of my rootkit as it was in at the end of the HACS408T Penetration Testing class at UMD. It is most definitely still a work in progress. There are many (semi) useless files laying around which I was using as sample code to test things as well. However, it should compile and be able to be loaded into the kernel. Developed targeting kernel versions around linux-4.x and 3.x, but also probably works (currently, anyway in this basic state) on kernel 2.x.